% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_birdcallbr.R
\name{birdcallbr_dataset}
\alias{birdcallbr_dataset}
\title{Birdcall Brazil Dataset}
\usage{
birdcallbr_dataset(
  root,
  audio_duration = 5,
  train = TRUE,
  url = "birdcallbr_v1",
  folder_in_archive = "BirdcallBR",
  download = FALSE
)
}
\arguments{
\item{root}{(str): Path to the directory where the dataset is found or downloaded.}

\item{audio_duration}{(dbl): Duration in seconds of the audio. Allowed values are 1, 2 or 5. (default: 5)}

\item{train}{(bool, optional): If True, creates dataset from `audio_ids_train.rds`, otherwise from `audio_ids_test.rds`.}

\item{url}{(str, optional): The URL to download the dataset from, or the type of the dataset to dowload.
Allowed type values are ``"birdcallbr_v1"`` (default: ``"birdcallbr_v1"``)}

\item{folder_in_archive}{(str, optional): The top-level directory of the dataset.  (default: ``"BirdcallBR"``)}

\item{download}{(bool, optional): Whether to download the dataset if it is not found at root path.  (default: ``FALSE``).}
}
\description{
Birdcall Brazil Dataset
}
