% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggpixelgrid}
\alias{ggpixelgrid}
\title{Plot Pixel Grids}
\usage{
ggpixelgrid(
  image_tensor_tbl,
  nrow = NULL,
  ncol = NULL,
  x = x,
  y = y,
  channels = dplyr::starts_with("c"),
  label = i,
  grid_tickness = 0,
  grid_colour = "grey",
  ...
)
}
\arguments{
\item{image_tensor_tbl}{image tensor tibble. Usually a output from [mestrado::image_tensors_to_tbl].}

\item{nrow}{facet_wrap's nrow.}

\item{ncol}{facet_wrap's ncol.}

\item{x}{bare name. Column of coordinate X.}

\item{y}{bare name. Column of coordinate Y.}

\item{channels}{<tidy-select> columns of channels. It must select 3 or less columns (for R, G and B). If less than 3, channels will be recycled.}

\item{label}{expr. Column or expression to put as title for each image.}

\item{grid_tickness}{numeric. Tickness of the pixel grid.}

\item{grid_colour}{character. Colour of the pixel grid.}

\item{...}{arguments passed to facet_wrap.}
}
\description{
Helper function to plot images from "image tensors" data.frames. Designed to work with [mestrado::image_tensors_to_tbl].
}
