% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{identity_resampling}
\alias{identity_resampling}
\title{see this issue https://github.com/rstudio/tensorflow/issues/402
sudo apt install sox}
\usage{
identity_resampling(orig, dest = orig)
}
\description{
see this issue https://github.com/rstudio/tensorflow/issues/402
sudo apt install sox
}
