% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{image_tensors_to_tbl}
\alias{image_tensors_to_tbl}
\title{Images Tensor to Coordinate data.frame}
\usage{
image_tensors_to_tbl(image_tensors)
}
\arguments{
\item{image_tensors}{tensor with dimension (B x C x H x W).}
}
\value{
data.frame with B*H*W rows and 3 + C columns.
}
\description{
Converts a batch of (image) torch tensors (B x C x H x W) into a data.frame with xy coordinates along with channels (x, y, c1, ..., cn).
Useful along with ggplot2 plotting.
}
\examples{

imgs <- torch::torch_tensor(1:prod(2*3*3*4))$reshape(c(2,3,3,4))
image_tensors_to_tbl(imgs)

}
