% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_labeling.R
\name{install_iranges}
\alias{install_iranges}
\title{Install IRange}
\usage{
install_iranges()
}
\value{
return nothing.
}
\description{
Helper to install IRange from Biocondutor.
}
\examples{
library(mestrado)
install_iranges()
}
