% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_pixel_matrix}
\alias{plot_pixel_matrix}
\title{plot_pixel_matrix}
\usage{
plot_pixel_matrix(pixel_matrix, title = "")
}
\value{

}
\description{
Plot a pixel matrix with magma palette. Made to plot spectrograms.
}
