% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_labeling.R
\name{retrieve_labels_for_one_dir}
\alias{retrieve_labels_for_one_dir}
\title{Retrieve Labels}
\usage{
retrieve_labels_for_one_dir(slices_dir, tidy_annotations, pattern = NULL)
}
\arguments{
\item{slices_dir}{atomic character. Path to dir where the slices are (see \code{\link{slice_wavs}}).}

\item{tidy_annotations}{a tibble with the annotations (see \code{\link{tidy_annotations}}).}

\item{pattern}{atomic character. A regex pattern to select files from 'slices_dir'. Useful to filter 
out just specifc types of labels (see examples).}
}
\value{

}
\description{
Plug the labels from a set of annotations (made by \code{\link[wavesurfer]{annotator_app}}) in 
a dataset of slices (made by \code{\link{slice_wavs}}).
}
\examples{

}
