% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_slicing.R
\name{slice_one_wav_sequencially}
\alias{slice_one_wav_sequencially}
\title{Slice one Wave File Sequentially}
\usage{
slice_one_wav_sequencially(
  audio_file,
  audio_dir,
  audio_dir_dest,
  interval = 1,
  overlap = 0,
  invisible = FALSE
)
}
\arguments{
\item{audio_file}{}

\item{audio_dir}{}

\item{audio_dir_dest}{}

\item{interval}{}

\item{overlap}{}

\item{invisible}{}
}
\value{
one character with the directory where the slices were stored.
}
\description{
Slice one wave file sequentially by regular interval of time with ou without overlap.
}
