% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-slicing.R
\name{slice_wavs}
\alias{slice_wavs}
\title{Slice a Set of Wave Files}
\usage{
slice_wavs(audio_dir, audio_dir_dest = NULL, interval = 1, parallel = 1)
}
\arguments{
\item{audio_dir}{}

\item{audio_dir_dest}{}

\item{interval}{}

\item{parallel}{}

\item{audio_ext}{}
}
\value{

}
\description{
Vectorized version of slice_one_wav_*() family of functions. Store in disk 
the slices of every file of a vector of mp3 file names.
}
