% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_labeling.R
\name{tidy_annotations}
\alias{tidy_annotations}
\title{Tidy Annotations data.frame}
\usage{
tidy_annotations(annotations_dir)
}
\arguments{
\item{annotations_dir}{atomic character. Path to the annotations directory.}
}
\value{
tibble which each row is an annotation.
}
\description{
Build the data.frame from a directory with the annotations made with \code{\link[wavesurfer]{annotator_app()}}.
}
\examples{
library(mestrado)
annotation_dir <-  system.file("annotations", package = "mestrado")
annotation_tibble <- tidy_annotations(annotation_dir)
}
