% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{tidy_audio}
\alias{tidy_audio}
\title{Tidy MP3 or Wave}
\usage{
tidy_audio(wav, sample_rate = 16000, bits = 16)
}
\arguments{
\item{wav}{Wave object. Usualy returned by \code{\link[tuneR]{readMP3}}/\code{\link[tuneR]{readWave}}.}
}
\value{
a transformed Wave object.
}
\description{
Apply transformations on an wav object for modelling purposes.
}
\details{
The transformations are:

- mp3 to wav
- normalization
- stereo to mono
}
\examples{

library(tuneR)
mp3_dir <- system.file("mp3_sample", package = "mestrado")
mp3_files <- list.files(mp3_dir, full.names = TRUE)
wav <- readMP3(mp3_files[1])

transformed_wav <- tidy_audio(wav)
}
