% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_slicing.R
\name{slice_wavs}
\alias{slice_wavs}
\title{Slice a Set of Wave Files}
\usage{
slice_wavs(audio_dir, audio_dir_dest = NULL, interval = 1, parallel = 1)
}
\arguments{
\item{audio_dir}{}

\item{audio_dir_dest}{}

\item{interval}{}

\item{parallel}{}

\item{audio_ext}{}
}
\value{

}
\description{
Vectorized version of slice_one_wav_*() family of functions. Store in disk 
the slices of every file of a vector of mp3 file names.
}
\examples{

library(mestrado)

wav_dir <- system.file("wav_sample", package = "mestrado")
temp_dir <- tempdir()

slices_path <- slice_wavs(wav_dir, temp_dir)
slices_path

slices <- list.files(slices_path)
slices[4:7]
}
